+++ Assets/PositionBasedDynamics/Scripts/Collisions/CollisionContact3d.cs
@@ -11,7 +11,7 @@ namespace PositionBasedDynamics.Collisions
     internal abstract class CollisionContact3d
     {
 
-        internal abstract void ResolveContact(double di);
+        internal abstract void ResolveContact(float di);
 
     }
